+++ CHANGELOG.md
@@ -11,6 +11,8 @@ FlatLaf Change Log
 
 #### Fixed bugs
 
+- CheckBox and RadioButton: Fixed styling of custom icon. Also fixed focus width
+  (and preferred size) if using custom icon. (PR #1060)
 - TextField: Fixed wrong leading/trailing icon placement if border is set to
   `null`. (issue #1047)
 - Extras: UI defaults inspector: Exclude inspector window from being blocked by